% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.assert_pos_scale}
\alias{.assert_pos_scale}
\title{Validate a transform scale}
\usage{
.assert_pos_scale(scale_c)
}
\arguments{
\item{scale_c}{Candidate scale \code{c}.}
}
\description{
Validate a transform scale
}
\keyword{internal}
