% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.build_imputed_df}
\alias{.build_imputed_df}
\title{Merge posterior estimates back into the original long-format df}
\usage{
.build_imputed_df(df, post_long, impute_kind, drivers, return = "point")
}
\description{
Merge posterior estimates back into the original long-format df
}
\keyword{internal}
