% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{.DOC_LIKE_ANALYTES}
\alias{.DOC_LIKE_ANALYTES}
\title{Analytes that satisfy the organic-carbon hurdle for the organics model}
\format{
An object of class \code{character} of length 5.
}
\usage{
.DOC_LIKE_ANALYTES
}
\description{
Analytes that satisfy the organic-carbon hurdle for the organics model
}
\keyword{internal}
