% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.fit_group_model}
\alias{.fit_group_model}
\title{Fit a single brms group model (metals or organics)}
\usage{
.fit_group_model(
  df,
  target_analytes,
  pca_obj,
  family,
  iter,
  warmup,
  chains,
  cores,
  impute_method = "rescor_mi",
  ...
)
}
\description{
Fit a single brms group model (metals or organics)
}
\keyword{internal}
