% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_model.R
\name{.hydro_pred}
\alias{.hydro_pred}
\title{Deterministic season-blind hydrology prediction \code{beta.f(hydro)} at query dates}
\usage{
.hydro_pred(m, feats, qdates)
}
\description{
Deterministic season-blind hydrology prediction \code{beta.f(hydro)} at query dates
}
\keyword{internal}
