% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{.METAL_ANALYTES}
\alias{.METAL_ANALYTES}
\title{All metal-type analytes (used to define the metals imputation group)}
\format{
An object of class \code{character} of length 22.
}
\usage{
.METAL_ANALYTES
}
\description{
All metal-type analytes (used to define the metals imputation group)
}
\keyword{internal}
