% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{.nh3_unionised_fraction}
\alias{.nh3_unionised_fraction}
\title{Un-ionised ammonia fraction (NH3) of total ammonia-N}
\usage{
.nh3_unionised_fraction(pH, temperature_C)
}
\arguments{
\item{pH}{Sample pH.}

\item{temperature_C}{Sample temperature (°C).}
}
\value{
Fraction (0–1) of total ammonia-N present as un-ionised NH3, using
\code{f = 1 / (1 + 10^(pKa - pH))} with \code{pKa = 0.09018 + 2729.92 / T(K)}
(Emerson et al. 1975), \code{T(K) = temperature_C + 273.15}.
}
\description{
Un-ionised ammonia fraction (NH3) of total ammonia-N
}
\keyword{internal}
