% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.predict_and_merge}
\alias{.predict_and_merge}
\title{Predict and merge imputed values for one analyte group}
\usage{
.predict_and_merge(df, group, pca_scores, eligible_ids, return)
}
\description{
Predict and merge imputed values for one analyte group
}
\keyword{internal}
