% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.ref_summary}
\alias{.ref_summary}
\title{Compute a reference-distribution summary statistic}
\usage{
.ref_summary(x, summary, eps = 1e-09)
}
\description{
Compute a reference-distribution summary statistic
}
\keyword{internal}
