% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.ref_summary_bootstrap_ci}
\alias{.ref_summary_bootstrap_ci}
\title{Bootstrap CI for a reference summary statistic}
\usage{
.ref_summary_bootstrap_ci(x, summary, eps = 1e-09, n_boot = 1000L)
}
\description{
Bootstrap CI for a reference summary statistic
}
\keyword{internal}
