% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.reshape_posterior_draws}
\alias{.reshape_posterior_draws}
\title{Reshape brms::posterior_predict() output to long format}
\usage{
.reshape_posterior_draws(post_draws, sample_ids, safe_analytes)
}
\description{
Reshape brms::posterior_predict() output to long format
}
\keyword{internal}
