% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{.DOC_LIKE_ANALYTES}
\alias{.DOC_LIKE_ANALYTES}
\title{Analytes that satisfy the organic-carbon hurdle for the organics group}
\format{
An object of class \code{character} of length 5.
}
\usage{
.DOC_LIKE_ANALYTES
}
\description{
A leachate-preset default: the presence hurdle for the \code{organics}
\code{\link[=impute_group]{impute_group()}} in \code{\link[=leachate_impute_groups]{leachate_impute_groups()}}.
}
\keyword{internal}
