% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{.IMPUTE_EXCLUDED}
\alias{.IMPUTE_EXCLUDED}
\title{Analytes that must never enter the imputation model as response variables}
\format{
An object of class \code{character} of length 10.
}
\usage{
.IMPUTE_EXCLUDED
}
\description{
These are excluded from both the metals and organics groups.  They are
typically non-concentration measurements (counts, qualitative, physical)
for which a log-normal concentration model is inappropriate.
}
\keyword{internal}
