% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{.reshape_posterior_means}
\alias{.reshape_posterior_means}
\title{Reshape brms::fitted() output to long format}
\usage{
.reshape_posterior_means(post_means, sample_uids, safe_analytes)
}
\description{
Reshape brms::fitted() output to long format
}
\keyword{internal}
