% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\docType{data}
\name{.WQ_BLOCK_CANDIDATES}
\alias{.WQ_BLOCK_CANDIDATES}
\title{WQ analytes used in the PCA pre-processing step}
\format{
An object of class \code{character} of length 32.
}
\usage{
.WQ_BLOCK_CANDIDATES
}
\description{
All candidates; the actual variables used at any given site are the
intersection of this list with analytes that pass prescreen in the training
data.  ORP and DO are included because they are valuable at sites where they
are measured, but at most sites they will be filtered out by prescreen.
}
\details{
Note: SO4 uses \verb{\\u00b2\\u207b} (\u00b2\u207b) escape sequences to ensure correct
string matching regardless of how the source file is parsed by devtools.
}
\keyword{internal}
