% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leachatetools-package.R
\docType{package}
\name{leachatetools-package}
\alias{leachatetools}
\alias{leachatetools-package}
\title{leachatetools: Water Quality Analysis Tools for Leachate-Impacted Waters}
\description{
Tools for analysing water quality in systems potentially impacted by landfill leachate. Core capabilities: species sensitivity distribution (SSD) based potentially affected fraction (PAF) calculations using ANZECC/ANZG toxicity data; multi-substance PAF (msPAF / AmsPAF) with local background adjustment via the Added Risk Approach; and end-member mixing analysis for leachate detection (LMF).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.kedumba.com.au/leachatetools/}
  \item \url{https://github.com/vorpalvorpal/leachatetools}
  \item Report bugs at \url{https://github.com/vorpalvorpal/leachatetools/issues}
}

}
\author{
\strong{Maintainer}: Robin Shannon \email{rjs@robinshannon.net}

}
\keyword{internal}
